+++ wagtail_localize/translation/__init__.py
@@ -1,0 +1,3 @@
+default_app_config = (
+    "wagtail_localize.translation.apps.WagtailLocalizeTranslationAppConfig"
+)
