+++ wagtail_localize/translation/apps.py
@@ -1,0 +1,7 @@
+from django.apps import AppConfig
+
+
+class WagtailLocalizeTranslationAppConfig(AppConfig):
+    label = "wagtail_localize_translation_memory"
+    name = "wagtail_localize.translation"
+    verbose_name = "Wagtail Localize Translation"
