--- wagtail_localize/translation_memory/__init__.py
@@ -1,3 +1,0 @@
-default_app_config = (
-    "wagtail_localize.translation_memory.apps.WagtailLocalizeTranslationMemoryAppConfig"
-)
