--- wagtail_localize/translation_memory/apps.py
@@ -1,7 +1,0 @@
-from django.apps import AppConfig
-
-
-class WagtailLocalizeTranslationMemoryAppConfig(AppConfig):
-    label = "wagtail_localize_translation_memory"
-    name = "wagtail_localize.translation_memory"
-    verbose_name = "Wagtail Localize Translation Memory"
